Updated descriptions for parameters
Updated various descriptions around the various parameters as well as a note about a specific parameter that has an open Bug - See Adam Meltzer for details of the bug.

I added a Note - about what is and what isn't supported in the current parameter usage rather than calling it out as a bug.