Clarify Get-CMSoftwareUpdateGroup CI_ID Usage
EXAMPLE #2 Uses an example ID of "ST10000D" (in the common CM Object ID format of <site code>+<six alphanumeric digits>), except that the 'Id' parameter is an Int32 type (no letters allowed). It is, in fact, looking for the Config Item ID of the Software Update Group (which is an integer). Additionally, added language in both the example and the Id parameter description to make it clear the parameter wants the Config Item ID (CI_ID) and not another ID type, such as the CI Unique ID.